refactor(tool): extract commands in separate files
